requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
ephem>=4.1.5
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
apscheduler>=3.10.0
pydantic>=2.7.0
